requirements.txt
# ONNX Runtime (GPU support)
onnxruntime-gpu>=1.20.1

# Web Framework
flask>=3.0.0
flask-cors>=4.0.0
werkzeug>=3.0.0
gunicorn>=20.1.0

# HTTP Client
requests>=2.32.0

# Image Processing
Pillow>=10.0.0
opencv-python>=4.12.0.88

# NumPy (required by ONNX Runtime)
numpy>=1.24.0

# PyTorch (for TorchScript model support)
torch>=2.0.0

# Utilities
rich>=14.0.0
